Caching getSystemInfoSystem information in mini game platforms
